Se agrego un endpoint para obtener un usuario por medio de su correo y por su id.
Se modifico el endpoint de /ingresarDireccion, para que reciba el id_usuario y devuelva las direcciones asignadas a ese usuario.

Se actualizo en el Modelo Libro, el tipo de dato para Precio e Impuesto. Ambos pasaron de Int a Double.

Todos los endpoint de direccion piden el token de acceso para: eliminar, modificar, insertar u obtener datos.

Los endpoint del LibroController: /crearLibro, /libro/{id} se modificaron para que se requiera un token de acceso para acceder a ellos.

Los endpoint de UsuarioController: /update, listUser y /UserByEmail se modificaron para que requieran el token de acceso para acceder a ellos.

+++ Biblioteca/src/main/java/com/CRUD/Biblioteca/Model/Libro.java
@@ -28,10 +28,10 @@ public class Libro {
     private String portada;
 
     @Column(nullable = false)
-    private int precio;
+    private double precio;
 
     @Column(nullable = false)
-    private int impuesto;
+    private double impuesto;
 
     @Column(nullable = false)
     private int stock;
@@ -59,7 +59,7 @@ public class Libro {
 
     public Libro() {super();}
 
-    public Libro(Integer id, GeneroLiterario generoLiterario, String nombre, String sinopsis, String id_portada, String portada, int precio, int impuesto, int stock, String autor, String editorial, String edicion, Date fecha_publicacion, String idioma, int numero_paginas, String dimensiones, String coleccion) {
+    public Libro(Integer id, GeneroLiterario generoLiterario, String nombre, String sinopsis, String id_portada, String portada, double precio, double impuesto, int stock, String autor, String editorial, String edicion, Date fecha_publicacion, String idioma, int numero_paginas, String dimensiones, String coleccion) {
         this.id = id;
         this.generoLiterario = generoLiterario;
         this.nombre = nombre;
@@ -127,19 +127,19 @@ public class Libro {
         this.portada = portada;
     }
 
-    public int getPrecio() {
+    public double getPrecio() {
         return precio;
     }
 
-    public void setPrecio(int precio) {
+    public void setPrecio(double precio) {
         this.precio = precio;
     }
 
-    public int getImpuesto() {
+    public double getImpuesto() {
         return impuesto;
     }
 
-    public void setImpuesto(int impuesto) {
+    public void setImpuesto(double impuesto) {
         this.impuesto = impuesto;
     }
